refactor shared WS logic into useWebsocket hook
