+++ awx/ui_next/src/components/JobList/useWsJobs.js
@@ -1,5 +1,6 @@
-import { useState, useEffect, useRef } from 'react';
+import { useState, useEffect } from 'react';
 import { useLocation } from 'react-router-dom';
+import useWebsocket from '../../util/useWebsocket';
 import useThrottle from '../../util/useThrottle';
 import { parseQueryString } from '../../util/qs';
 import sortJobs from './sortJobs';
@@ -7,9 +8,13 @@ import sortJobs from './sortJobs';
 export default function useWsJobs(initialJobs, fetchJobsById, qsConfig) {
   const location = useLocation();
   const [jobs, setJobs] = useState(initialJobs);
-  const [lastMessage, setLastMessage] = useState(null);
   const [jobsToFetch, setJobsToFetch] = useState([]);
   const throttledJobsToFetch = useThrottle(jobsToFetch, 5000);
+  const lastMessage = useWebsocket({
+    jobs: ['status_changed'],
+    schedules: ['changed'],
+    control: ['limit_reached_1'],
+  });
 
   useEffect(() => {
     setJobs(initialJobs);
@@ -37,8 +42,6 @@ export default function useWsJobs(initialJobs, fetchJobsById, qsConfig) {
     })();
   }, [throttledJobsToFetch, fetchJobsById]); // eslint-disable-line react-hooks/exhaustive-deps
 
-  const ws = useRef(null);
-
   useEffect(() => {
     if (!lastMessage || !lastMessage.unified_job_id) {
       return;
@@ -60,51 +63,6 @@ export default function useWsJobs(initialJobs, fetchJobsById, qsConfig) {
       enqueueJobId(lastMessage.unified_job_id);
     }
   }, [lastMessage]); // eslint-disable-line react-hooks/exhaustive-deps
-
-  useEffect(() => {
-    ws.current = new WebSocket(`wss://${window.location.host}/websocket/`);
-
-    const connect = () => {
-      const xrftoken = `; ${document.cookie}`
-        .split('; csrftoken=')
-        .pop()
-        .split(';')
-        .shift();
-      ws.current.send(
-        JSON.stringify({
-          xrftoken,
-          groups: {
-            jobs: ['status_changed'],
-            schedules: ['changed'],
-            control: ['limit_reached_1'],
-          },
-        })
-      );
-    };
-    ws.current.onopen = connect;
-
-    ws.current.onmessage = e => {
-      setLastMessage(JSON.parse(e.data));
-    };
-
-    ws.current.onclose = e => {
-      // eslint-disable-next-line no-console
-      console.debug('Socket closed. Reconnecting...', e);
-      setTimeout(() => {
-        connect();
-      }, 1000);
-    };
-
-    ws.current.onerror = err => {
-      // eslint-disable-next-line no-console
-      console.debug('Socket error: ', err, 'Disconnecting...');
-      ws.current.close();
-    };
-
-    return () => {
-      ws.current.close();
-    };
-  }, []);
 
   return jobs;
 }
