+++ awx/ui_next/src/screens/Inventory/InventoryList/useWsInventories.js
@@ -1,4 +1,5 @@
-import { useState, useEffect, useRef } from 'react';
+import { useState, useEffect } from 'react';
+import useWebsocket from '../../../util/useWebsocket';
 import useThrottle from '../../../util/useThrottle';
 
 export default function useWsProjects(
@@ -6,10 +7,13 @@ export default function useWsProjects(
   fetchInventoriesById
 ) {
   const [inventories, setInventories] = useState(initialInventories);
-  const [lastMessage, setLastMessage] = useState(null);
   const [inventoriesToFetch, setInventoriesToFetch] = useState([]);
   const throttledInventoriesToFetch = useThrottle(inventoriesToFetch, 5000);
-  const ws = useRef(null);
+  const lastMessage = useWebsocket({
+    inventories: ['status_changed'],
+    jobs: ['status_changed'],
+    control: ['limit_reached_1'],
+  });
 
   useEffect(() => {
     setInventories(initialInventories);
@@ -83,51 +87,6 @@ export default function useWsProjects(
     // eslint-disable-next-line react-hooks/exhaustive-deps,
     [lastMessage]
   );
-
-  useEffect(() => {
-    ws.current = new WebSocket(`wss://${window.location.host}/websocket/`);
-
-    const connect = () => {
-      const xrftoken = `; ${document.cookie}`
-        .split('; csrftoken=')
-        .pop()
-        .split(';')
-        .shift();
-      ws.current.send(
-        JSON.stringify({
-          xrftoken,
-          groups: {
-            inventories: ['status_changed'],
-            jobs: ['status_changed'],
-            control: ['limit_reached_1'],
-          },
-        })
-      );
-    };
-    ws.current.onopen = connect;
-
-    ws.current.onmessage = e => {
-      setLastMessage(JSON.parse(e.data));
-    };
-
-    ws.current.onclose = e => {
-      // eslint-disable-next-line no-console
-      console.debug('Socket closed. Reconnecting...', e);
-      setTimeout(() => {
-        connect();
-      }, 1000);
-    };
-
-    ws.current.onerror = err => {
-      // eslint-disable-next-line no-console
-      console.debug('Socket error: ', err, 'Disconnecting...');
-      ws.current.close();
-    };
-
-    return () => {
-      ws.current.close();
-    };
-  }, []);
 
   return inventories;
 }
